restructured how Ehcache works within the SessionManager, specifically to address default activeSessionCache settings.  This removed the need for the jsecurity-specific failsafe xml file.
git-svn-id: https://svn.apache.org/repos/asf/incubator/jsecurity/trunk@710522 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/jsecurity/session/support/eis/SessionDAO.java
@@ -111,5 +111,5 @@ id.equals( session.getId() ) == true</pre></code>
      * reporting.
      * @return the total number of active sessions (those that haven't been stopped or expired).
      */
-    int getActiveSessionCount();
+    long getActiveSessionCount();
 }
